+++ src/lib.rs
@@ -80,6 +80,7 @@
 pub mod container;
 pub mod decoder;
 pub mod header;
+pub mod it;
 pub mod mixer;
 pub mod player;
 pub mod samples;